mm, debug: introduce dump_gfpflag_names() for symbolic printing of gfp_flags
It would be useful to convert gfp_flags into string representation when
printing them in case of allocation failure, OOM etc.  There's a script
./scripts/gfp-translate to make this simpler, but it needs the matching
version of the sources to be accurate, and the flags have been undergoing
some changes recently.

The ftrace framework already has this translation in the form of
show_gfp_flags() defined in include/trace/events/gfpflags.h which defines
the translation table internally.  Allow reusing the table outside ftrace
by putting it behind __def_gfpflag_names definition and introduce
dump_gfpflag_names() to handle the printing.

While at it, also fill in the names for the flags and flag combinations
that have been missing in the table.  GFP_NOWAIT no longer equals to "no
flags", so change the output for no flags to "none".

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Sasha Levin <sasha.levin@oracle.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>