printk/nmi: Remove the questionable CONFIG_NEED_PRINTK_NMI
The flag NEED_PRINTK_NMI was added because of Arm.  It used the NMI safe
backtrace implementation on all Arm systems.  But it did not have a real
NMI handling on CPU_V7M.

It seems that it causes more confusion than good.  Let's use HAVE_NMI on
all arm systems and get rid of the problematic flag.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>