ubsan-run-time-undefined-behavior-sanity-checker-fix-5
Don't use 'bounds-strict' checker. It's a new checker that showed up
in GCC 6. It appeared to cause a lot of false positives in dynamically
allocated structs which has array as the last field, e.g:
	[   72.707515] UBSAN: Undefined behaviour in kernel/pid.c:506:23
	[   72.707517] index 2 is out of range for type 'upid [1]'

Anyway, we have KASAN for these kind of bugs, so we won't lose much by
disabling bounds-strict.

Signed-off-by: Andrey Ryabinin <aryabinin@virtuozzo.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>