watchdog: don't run proc_watchdog_update if new value is same as old
While working on a script to restore all sysctl params before a series of
tests I found that writing any value into the
/proc/sys/kernel/{nmi_watchdog,soft_watchdog,watchdog,watchdog_thresh}
causes them to call proc_watchdog_update().  Not only that, but when I
wrote to these proc files in a loop I could easily trigger a soft lockup.

[  955.756196] NMI watchdog: enabled on all CPUs, permanently consumes one hw-PMU counter.
[  955.765994] NMI watchdog: enabled on all CPUs, permanently consumes one hw-PMU counter.
[  955.774619] NMI watchdog: enabled on all CPUs, permanently consumes one hw-PMU counter.
[  955.783182] NMI watchdog: enabled on all CPUs, permanently consumes one hw-PMU counter.
[  959.788319] NMI watchdog: BUG: soft lockup - CPU#4 stuck for 30s! [swapper/4:0]
[  959.788325] NMI watchdog: BUG: soft lockup - CPU#5 stuck for 30s! [swapper/5:0]

There doesn't appear to be a reason for doing this work other every time a
write occurs, so only do the work when the values change.

Signed-off-by: Josh Hunt <johunt@akamai.com>
Cc: Ulrich Obergfell <uobergfe@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>