m68k: rename UL() to TO_UL()
Patch series "linux/const.h: cleanups of macros such as UL(), _BITUL(), BIT() etc", v2.

ARM, ARM64, UniCore32 define UL() as a shorthand of _AC(..., UL).  In the
future, more architectures may introduce it, so move the definition to
include/linux/const.h.

The _AC() is used for either (likely) UL or (unlikely) ULL.  Having UL(L)
in a common place can avoid direct use of _AC().

The _AC() is defined under the uapi directory, so it compels
underscore-prefixed macros even for unexported headers.

I see similar situation for _BITUL().  This is available in both C and
assembly.  However, it is defined in an uapi header, so direct use of the
underscore macro is needed even for unexported headers.  The 3/3 makes
BIT() available in assembly too, which will be more suitable for use in
unexported headers.


This patch (of 4):

ARM, ARM64 and UniCore32 define UL(x) like follows:

  #define UL(x) _AC(x, UL)

While, M68K defines it differently:
  #define UL(x) ((unsigned long) (x))

I am moving the former to a common header in the next commit.
Beforehand, this commit renames the latter to avoid name conflict.

Link: http://lkml.kernel.org/r/1483582810-7046-2-git-send-email-yamada.masahiro@socionext.com
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Acked-by: Geert Uytterhoeven <geert@linux-m68k.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>