memory_hotplug: zone_can_shift() returns boolean value
online_{kernel|movable} is used to change the memory zone to
ZONE_{NORMAL|MOVABLE} and online the memory.

To check that memory zone can be changed, zone_can_shift() is used. 
Currently the function returns minus integer value, plus integer value and
0.  When the function returns minus or plus integer value, it means that
the memory zone can be changed to ZONE_{NORNAL|MOVABLE}.

But when the function returns 0, there is 2 meanings.

One of the meanings is that the memory zone does not need to be changed. 
For example, when memory is in ZONE_NORMAL and onlined by online_kernel
the memory zone does not need to be changed.

Another meaning is that the memory zone cannot be changed.  When memory is
in ZONE_NORMAL and onlined by online_movable, the memory zone may not be
changed to ZONE_MOVALBE due to memory online limitation(see
Documentation/memory-hotplug.txt).  In this case, memory must not be
onlined.

The patch changes the return type of zone_can_shift() so that memory is
not onlined when memory zone cannot be changed.

Fixes: df429ac03936 ("memory-hotplug: more general validation of zone during online")
Link: http://lkml.kernel.org/r/2f9c3837-33d7-b6e5-59c0-6ca4372b2d84@gmail.com
Signed-off-by: Yasuaki Ishimatsu <isimatu.yasuaki@jp.fujitsu.com>
Reviewed-by: Reza Arbab <arbab@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>